requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
APScheduler==3.10.4
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
frozenlist==1.7.0
greenlet==3.2.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
multidict==6.5.1
outcome==1.3.0.post0
playwright==1.53.0
propcache==0.3.2
pycparser==2.22
pyee==13.0.0
PySocks==1.7.1
python-telegram-bot==20.3
pytz==2025.2
requests==2.32.4
selenium==4.33.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
